requirements.txt
pystray
pillow
requests
threading
time